backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
youtube-transcript-api
PyPDF2
numpy
scikit-learn
groq
langchain-text-splitters
python-dotenv
pydantic
